R: move test utilities into subfolder
