+++ Microsoft.DotNet.Interactive.ExtensionLab.Tests/DownloadTests.cs
@@ -8,6 +8,7 @@ using FluentAssertions;
 using Microsoft.DotNet.Interactive.CSharp;
 using Microsoft.DotNet.Interactive.Events;
 using Microsoft.DotNet.Interactive.Tests;
+using Microsoft.DotNet.Interactive.Tests.Utility;
 using Xunit;
 
 namespace Microsoft.DotNet.Interactive.ExtensionLab.Tests
