+++ Microsoft.DotNet.Interactive.ExtensionLab.Tests/TranscriptTests.cs
@@ -9,6 +9,7 @@ using System.Reactive.Disposables;
 using System.Threading.Tasks;
 using Microsoft.DotNet.Interactive.FSharp;
 using Microsoft.DotNet.Interactive.Tests;
+using Microsoft.DotNet.Interactive.Tests.Utility;
 using Xunit;
 
 namespace Microsoft.DotNet.Interactive.ExtensionLab.Tests
