+++ Microsoft.DotNet.Interactive.Tests/KernelBaseTests.cs
@@ -8,6 +8,7 @@ using FluentAssertions;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.DotNet.Interactive.Commands;
+using Microsoft.DotNet.Interactive.Tests.Utility;
 using Xunit;
 using Xunit.Abstractions;
