+++ Microsoft.DotNet.Interactive.Tests/LanguageKernelPackageTests.cs
@@ -12,6 +12,7 @@ using Microsoft.DotNet.Interactive.CSharp;
 using Microsoft.DotNet.Interactive.Events;
 using Microsoft.DotNet.Interactive.FSharp;
 using Microsoft.DotNet.Interactive.Jupyter;
+using Microsoft.DotNet.Interactive.Tests.Utility;
 using Newtonsoft.Json;
 using Recipes;
 using Xunit;
