+++ Microsoft.DotNet.Interactive.Tests/LanguageKernelTestBase.cs
@@ -10,6 +10,7 @@ using Microsoft.DotNet.Interactive.Events;
 using Microsoft.DotNet.Interactive.FSharp;
 using Microsoft.DotNet.Interactive.Jupyter;
 using Microsoft.DotNet.Interactive.PowerShell;
+using Microsoft.DotNet.Interactive.Tests.Utility;
 using Pocket;
 using Recipes;
 using Xunit.Abstractions;
