+++ Microsoft.DotNet.Interactive.Tests/LanguageKernelTests.cs
@@ -11,6 +11,7 @@ using FluentAssertions.Extensions;
 using Microsoft.DotNet.Interactive.Commands;
 using Microsoft.DotNet.Interactive.CSharp;
 using Microsoft.DotNet.Interactive.Events;
+using Microsoft.DotNet.Interactive.Tests.Utility;
 using XPlot.Plotly;
 using Xunit;
 using Xunit.Abstractions;
