+++ Microsoft.DotNet.Interactive.Tests/Server/StandardIOKernelServerTests.cs
@@ -13,6 +13,7 @@ using Microsoft.DotNet.Interactive.CSharp;
 using Microsoft.DotNet.Interactive.Events;
 using Microsoft.DotNet.Interactive.Jupyter;
 using Microsoft.DotNet.Interactive.Server;
+using Microsoft.DotNet.Interactive.Tests.Utility;
 using Pocket;
 using Xunit;
 using Xunit.Abstractions;
