+++ Microsoft.DotNet.Interactive.Tests/Utility/FakeKernel.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 using Microsoft.DotNet.Interactive.Commands;
 using Microsoft.DotNet.Interactive.Utility;
 
-namespace Microsoft.DotNet.Interactive.Tests
+namespace Microsoft.DotNet.Interactive.Tests.Utility
 {
     public class FakeKernel : KernelBase
     {
