+++ Microsoft.DotNet.Interactive.Tests/Utility/KernelExtensionTestHelper.cs
@@ -8,7 +8,7 @@ using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using Microsoft.DotNet.Interactive.Utility;
 
-namespace Microsoft.DotNet.Interactive.Tests
+namespace Microsoft.DotNet.Interactive.Tests.Utility
 {
     internal static class KernelExtensionTestHelper
     {
